+++ ayame/_typing.py
@@ -1,17 +1,21 @@
 #
 # ayame._typing
 #
-#   Copyright (c) 2025 Akinori Hattori <hattya@gmail.com>
+#   Copyright (c) 2025-2026 Akinori Hattori <hattya@gmail.com>
 #
 #   SPDX-License-Identifier: MIT
 #
 
+from collections.abc import Iterable
 import sys
-from typing import TYPE_CHECKING, Any, TypeAlias
+from typing import TYPE_CHECKING, runtime_checkable, Any, Protocol, TypeAlias, TypeVar
 
 
-__all__ = ['Headers', 'Locale', 'OptExcInfo', 'Self',
+__all__ = ['Headers', 'Locale', 'OptExcInfo', 'Self', 'SupportsKeysAndGetItem',
            'InputStream', 'StartResponse', 'WSGIEnvironment']
+
+_KT = TypeVar('_KT')
+_VT_co = TypeVar('_VT_co', covariant=True)
 
 Headers: TypeAlias = list[tuple[str, str]]
 Locale: TypeAlias = tuple[str | None, str | None]
@@ -30,3 +34,10 @@ elif TYPE_CHECKING:
 else:
     Self = Any
     InputStream = StartResponse = WSGIEnvironment = Any
+
+
+@runtime_checkable
+class SupportsKeysAndGetItem(Protocol[_KT, _VT_co]):
+
+    def keys(self) -> Iterable[_KT]: ...
+    def __getitem__(self, key: _KT, /) -> _VT_co: ...